add bottom padding in open-in-system-file-manager popup
