--- shared/fs/common/open-in-system-file-manager-container.desktop.js
@@ -1,21 +1,0 @@
-// @flow
-import * as FsGen from '../../actions/fs-gen'
-import {namedConnect} from '../../util/container'
-import {type OpenInSystemFileManagerProps as OwnProps} from './open-in-system-file-manager-container'
-import OpenInSystemFileManager from './open-in-system-file-manager.desktop'
-
-const mapStateToProps = (state, {path}: OwnProps) => ({
-  kbfsEnabled: state.fs.fileUI.driverStatus.type === 'enable',
-})
-
-const mapDispatchToProps = (dispatch, {path}: OwnProps) => ({
-  installFuse: () => dispatch(FsGen.createDriverEnable({})),
-  openInSystemFileManager: () => dispatch(FsGen.createOpenPathInSystemFileManager({path})),
-})
-
-export default namedConnect<OwnProps, _, _, _, _>(
-  mapStateToProps,
-  mapDispatchToProps,
-  (s, d, o) => ({...o, ...s, ...d}),
-  'ConnectedOpenInSystemFileManager'
-)(OpenInSystemFileManager)
