--- shared/fs/common/open-in-system-file-manager-container.native.js
@@ -1,2 +1,0 @@
-// @flow
-export default () => null
